requirements.txt
autopep8==1.5.4
py4j==0.10.9
pycodestyle==2.6.0
pyspark==3.0.1
requests==2.24.0
isodate==0.6.0
PyYAML~=5.3.1
setuptools~=47.1.0
numpy==1.19.3
pandas==1.1.3
